+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Module [INTERMIX](./docs/Intermix.md)
     - Added matrix mapping parameters on rows and columns for use with midi-mapping
     - Added option for excluding attenuverters from scenes
+    - Added ability to copy scenes
     - Fixed fading if either fade-in or fade-out is set to zero
 - Added dark panels for all modules (#15)
 - Added globals settings, esp. for dark panels als default